// backend/components/css/xiantiao/types.ts
export interface XiantiaoParams {
  cardBackground: string
  cardOpacity: number
  beforeCircleColor: string
  beforeCircleShadowColor: string
  beforeCircleSize: number
  beforeCircleTop: number
  beforeCircleRight: number
  afterCircleColor: string
  afterCircleSize: number
  afterCircleTop: number
  afterCircleRight: number
  afterCircleBorderWidth: number
  blurAmount: number
} 